+++ pgcli/main.py
@@ -205,16 +205,41 @@ class PGCli(object):
                                 'Execute commands from file.')
         self.pgspecial.register(self.write_to_file, '\\o', '\\o [filename]',
                                 'Send all query results to file.')
+        self.pgspecial.register(self.info_connection, '\\conninfo',
+                                '\\conninfo', 'Get connection details')
+
+    def info_connection(self, **_):
+        if self.pgexecute.host.startswith('/'):
+            host = 'socket "%s"' % self.pgexecute.host
+        else:
+            host = 'host "%s"' % self.pgexecute.host
+
+        yield (None, None, None, 'You are connected to database "%s" as user '
+               '"%s" on %s at port "%s".' % (self.pgexecute.dbname,
+                                             self.pgexecute.user,
+                                             host,
+                                             self.pgexecute.port))
 
     def change_db(self, pattern, **_):
         if pattern:
-            db = pattern[1:-1] if pattern[0] == pattern[-1] == '"' else pattern
-            self.pgexecute.connect(database=db)
+            # Get all the parameters in pattern, handling double quotes if any.
+            infos = re.findall(r'"[^"]*"|[^"\'\s]+', pattern)
+            # Now removing quotes.
+            list(map(lambda s: s.strip('"'), infos))
+
+            infos.extend([None] * (4 - len(infos)))
+            db, user, host, port = infos
+            try:
+                self.pgexecute.connect(database=db, user=user, host=host,
+                                       port=port)
+            except OperationalError as e:
+                click.secho(str(e), err=True, fg='red')
+                click.echo("Previous connection kept")
         else:
             self.pgexecute.connect()
 
         yield (None, None, None, 'You are now connected to database "%s" as '
-                'user "%s"' % (self.pgexecute.dbname, self.pgexecute.user))
+               'user "%s"' % (self.pgexecute.dbname, self.pgexecute.user))
 
     def execute_from_file(self, pattern, **_):
         if not pattern: